extract public interface loader
