+++ apps/web-remix/app/routes/forms.$organizationId.pipelines.$pipelineId.tsx
@@ -1,5 +1,4 @@
 import React, { useMemo, useState } from 'react';
-import { json } from '@remix-run/node';
 import type { LoaderFunctionArgs } from '@remix-run/node';
 import { useLoaderData } from '@remix-run/react';
 import { withZod } from '@remix-validated-form/with-zod';
@@ -7,8 +6,6 @@ import { ValidatedForm } from 'remix-validated-form';
 import invariant from 'tiny-invariant';
 import { z } from 'zod';
 
-import type { IPipelinePublicResponse } from '~/api/pipeline/pipeline.contracts';
-import { PipelineApi } from '~/api/pipeline/PipelineApi';
 import { ChatMarkdown } from '~/components/chat/ChatMarkdown';
 import { Field } from '~/components/form/fields/field.context';
 import { FieldLabel } from '~/components/form/fields/field.label';
@@ -16,44 +13,20 @@ import { FieldMessage } from '~/components/form/fields/field.message';
 import { SmallFileInputField } from '~/components/form/fields/file.field';
 import { TextInputField } from '~/components/form/fields/text.field';
 import { SubmitButton } from '~/components/form/submit';
+import { publicInterfaceLoader } from '~/components/pages/pipelines/interface/interface.loader.server';
 import {
   NodeClearButton,
   NodeCopyButton,
   NodeDownloadButton,
 } from '~/components/pages/pipelines/Nodes/CustomNodes/NodeActionButtons';
 import { loaderBuilder } from '~/utils.server';
-import { UnauthorizedError } from '~/utils/errors';
-import type { ParsedResponse } from '~/utils/fetch.server';
 
 export async function loader(args: LoaderFunctionArgs) {
-  return loaderBuilder(async ({ request, params }, { fetch }) => {
+  return loaderBuilder(async ({ params, ...rest }, helpers) => {
     invariant(params.organizationId, 'organizationId not found');
     invariant(params.pipelineId, 'pipelineId not found');
 
-    const pipelineApi = new PipelineApi(fetch);
-
-    let pipeline: ParsedResponse<IPipelinePublicResponse> | void =
-      await pipelineApi
-        .getPipeline(params.organizationId, params.pipelineId)
-        .catch((e) => {
-          if (e instanceof UnauthorizedError) return;
-          throw e;
-        });
-    if (!pipeline) {
-      pipeline = await pipelineApi.getPublicPipeline(
-        params.organizationId,
-        params.pipelineId,
-      );
-    }
-
-    const alias = pipelineApi.getAliasFromUrl(request.url);
-
-    return json({
-      pipeline: pipeline.data,
-      organizationId: params.organizationId as string,
-      pipelineId: params.pipelineId as string,
-      alias,
-    });
+    return publicInterfaceLoader({ params, ...rest }, helpers);
   })(args);
 }
