+++ apps/web-remix/app/routes/voice-chat.$organizationId.pipelines.$pipelineId.tsx
@@ -1,49 +1,17 @@
 import React from 'react';
-import { json } from '@remix-run/node';
 import type { LoaderFunctionArgs } from '@remix-run/node';
 import { useLoaderData } from '@remix-run/react';
 import invariant from 'tiny-invariant';
 
-import type { IPipelinePublicResponse } from '~/api/pipeline/pipeline.contracts';
-import { PipelineApi } from '~/api/pipeline/PipelineApi';
-import { chatSize } from '~/components/chat/chat.types';
+import { publicInterfaceLoader } from '~/components/pages/pipelines/interface/interface.loader.server';
 import { loaderBuilder } from '~/utils.server';
-import { UnauthorizedError } from '~/utils/errors';
-import type { ParsedResponse } from '~/utils/fetch.server';
 
 export async function loader(args: LoaderFunctionArgs) {
-  return loaderBuilder(async ({ request, params }, { fetch }) => {
+  return loaderBuilder(async ({ params, ...rest }, helpers) => {
     invariant(params.organizationId, 'organizationId not found');
     invariant(params.pipelineId, 'pipelineId not found');
 
-    const pipelineApi = new PipelineApi(fetch);
-
-    let pipeline: ParsedResponse<IPipelinePublicResponse> | void =
-      await pipelineApi
-        .getPipeline(params.organizationId, params.pipelineId)
-        .catch((e) => {
-          if (e instanceof UnauthorizedError) return;
-          throw e;
-        });
-    if (!pipeline) {
-      pipeline = await pipelineApi.getPublicPipeline(
-        params.organizationId,
-        params.pipelineId,
-      );
-    }
-
-    const alias = pipelineApi.getAliasFromUrl(request.url);
-    const searchParams = new URLSearchParams(request.url);
-
-    const size = chatSize.safeParse(searchParams.get('size'));
-
-    return json({
-      pipeline: pipeline.data,
-      organizationId: params.organizationId as string,
-      pipelineId: params.pipelineId as string,
-      chatSize: size.success ? size.data : 'default',
-      alias,
-    });
+    return publicInterfaceLoader({ params, ...rest }, helpers);
   })(args);
 }
